Rename project to Luma.API and simplify structure
