--- Luma-API/Luma-API/Domain/Entities/Breakpoint.cs
@@ -1,2 +1,0 @@
-﻿namespace Luma_API.Domain.Entities;
-public enum Breakpoint { Lg, Md, Sm, Xs };
