--- Luma-API/Luma-API/Domain/Entities/Widget.cs
@@ -1,8 +1,0 @@
-﻿namespace Luma_API.Domain.Entities;
-
-public class Widget
-{
-    public Guid Id { get; set; }                          // PK
-    public string TypeKey { get; set; } = null!;          // e.g. "weather", "notes" (unique)
-    public string DisplayName { get; set; } = null!;
-}
