--- Luma-API/Luma-API/Domain/Entities/WidgetConfig.cs
@@ -1,12 +1,0 @@
-﻿using System.Text.Json;
-
-namespace Luma_API.Domain.Entities;
-public class WidgetConfig
-{
-    // PK = FK (1:1 with WidgetInstance)
-    public Guid WidgetInstanceId { get; set; }
-    public JsonDocument Config { get; set; } = JsonDocument.Parse("{}"); // JSONB in PG
-
-    // nav
-    public WidgetInstance WidgetInstance { get; set; } = null!;
-}
