--- Luma-API/Luma-API/Luma-API.http
@@ -1,6 +1,0 @@
-@Luma_API_HostAddress = http://localhost:5171
-
-GET {{Luma_API_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
