--- Luma-API/Luma-API/Models/WeatherDto.cs
@@ -1,9 +1,0 @@
-﻿namespace Luma_API.Models
-{
-    public class WeatherDto
-    {
-        public double TemperatureC { get; set; }
-        public string? Condition { get; set; }
-        public double Humidity { get; set; }
-    }
-}
