--- Luma-API/Luma-API/Services/IGeocodingService.cs
@@ -1,7 +1,0 @@
-﻿namespace Luma_API.Services
-{
-    public interface IGeocodingService
-    {
-        Task<(double lat, double lng)> GetCoordinatesAsync(string city);
-    }
-}
