--- Luma-API/Luma-API/Services/IWeatherService.cs
@@ -1,9 +1,0 @@
-﻿using Luma_API.Models;
-
-namespace Luma_API.Services
-{
-    public interface IWeatherService
-    {
-        Task<WeatherDto>GetWeatherAsync(double lat, double lng);
-    }
-}
