+++ Luma.API.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.14.36310.24 d17.14
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Luma.API", "Luma.API\Luma.API.csproj", "{8303F39C-13E2-4373-B419-A2AB31096F87}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8303F39C-13E2-4373-B419-A2AB31096F87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8303F39C-13E2-4373-B419-A2AB31096F87}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8303F39C-13E2-4373-B419-A2AB31096F87}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8303F39C-13E2-4373-B419-A2AB31096F87}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {0C144C2B-A70E-4CE9-8287-347B386C6BD9}
+	EndGlobalSection
+EndGlobal
