+++ Luma.API/Domain/Entities/Breakpoint.cs
@@ -1,0 +1,2 @@
+﻿namespace Luma.API.Domain.Entities;
+public enum Breakpoint { Lg, Md, Sm, Xs };
