+++ Luma.API/Luma.API.http
@@ -1,0 +1,6 @@
+@Luma.API_HostAddress = http://localhost:5171
+
+GET {{Luma.API_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
