+++ Luma.API/Models/WeatherDto.cs
@@ -1,0 +1,9 @@
+﻿namespace Luma.API.Models
+{
+    public class WeatherDto
+    {
+        public double TemperatureC { get; set; }
+        public string? Condition { get; set; }
+        public double Humidity { get; set; }
+    }
+}
