+++ Luma.API/Services/IGeocodingService.cs
@@ -1,0 +1,7 @@
+﻿namespace Luma.API.Services
+{
+    public interface IGeocodingService
+    {
+        Task<(double lat, double lng)> GetCoordinatesAsync(string city);
+    }
+}
