+++ Luma.API/Services/IWeatherService.cs
@@ -1,0 +1,9 @@
+﻿using Luma.API.Models;
+
+namespace Luma.API.Services
+{
+    public interface IWeatherService
+    {
+        Task<WeatherDto>GetWeatherAsync(double lat, double lng);
+    }
+}
